[CORE] Fix video glitches on video buffer overflow
Whenever there are no existing nodes to use, new node is trying to be allocated. That can happen only when the number of allocated nodes doesn't reach the current maximum video buffer size. Otherwise video buffer is forcibly reset. That causes visible, unnecessary video glitch.

With this fix, when the number of allocated nodes reach the current maximum video buffer size, the packets, which we know they won't be used anymore are discarded from the buffer. Nodes are released and one of them is given to currently processed packet straightaway, making the handling of buffer overflow very smooth.